removed warnings and turned on mipmaps

+++ Engine/src/Intersections.cpp
@@ -126,13 +126,13 @@ const bool Intersect::RayIntersectPlane(const Ray_t& ray, const comp::PlaneColli
 	float halfLenght = planeCollider.size.x / 2;
 
 	//Kollar ifall normalen och ray är parallela 
-	double dirNormal = direction.Dot(planeCollider.normal);
+	float dirNormal = direction.Dot(planeCollider.normal);
 	if (abs(dirNormal) < 0.0000001f)
 		return false;
 
 	//Plane intersecation test
 	sm::Vector3 originToCenter = planeCollider.center - origin;
-	double originToCenterDotNormal = originToCenter.Dot(planeCollider.normal);
+	float originToCenterDotNormal = originToCenter.Dot(planeCollider.normal);
 	t = originToCenterDotNormal / dirNormal;
 
 	////Return false if plane is behind us
@@ -142,7 +142,7 @@ const bool Intersect::RayIntersectPlane(const Ray_t& ray, const comp::PlaneColli
 	//Restriction that makes the plane to a disc, if its outside the disc radius it does not intersect
 	sm::Vector3 intersecPoint = origin + direction * (float)t;
 	sm::Vector3 disToCenter = (intersecPoint - planeCollider.center);
-	double disToCenterL = (intersecPoint - planeCollider.center).Length();
+	float disToCenterL = (intersecPoint - planeCollider.center).Length();
 
 	/*f (disToCenter.x >= halfLenght && disToCenter.y >= halfLenght)
 		return false;*/
